requirements.txt
flask
numpy
torch
torchvision
pillow